requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.6
botocore==1.34.6
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
decorator==5.1.1
docx==0.2.4
docx2txt==0.8
elevenlabs==0.2.26
emoji==2.10.1
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
Faker==21.0.0
fastapi==0.105.0
favicon==0.7.0
filelock==3.13.1
filetype==1.2.0
Flask==2.3.3
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
h11==0.14.0
htbuilder==0.6.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
install==1.3.5
ipdb==0.13.13
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.338
langdetect==1.0.9
langsmith==0.0.69
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==0.27.8
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pkg_resources==0.0.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.32.0
regex==2023.12.25
replicate==0.22.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.0
streamlit-antd-components==0.3.0
streamlit-authenticator==0.2.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-player==0.1.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.0.5
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.36.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.12.6
unstructured-client==0.18.0
urllib3==1.26.18
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
youtube-search-python==1.6.6
zipp==3.17.0
streamlit-antd-components==0.2.4
streamlit-player==0.1.5
streamlit-extras==0.2.7
google-search-results
